move cookie_helpers from /dashboard/lib to /lib Also introduce `CookieHelpers::SESSION_STORE` constant to extract session-store middleware configuration options for later reuse.
